# api-documentation.yaml
...
openapi: 3.0.0
info:
  description: This is the API for the Vehicle Cost Calculator
  version: "1.0.0"
  title: Vehicle Cost Calculator API
...
paths:
...
  /users:
      post:
        tags:
          - Users
        summary: Add a new User
        description: |
          Adds a new User to the system
        requestBody:
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/UserInput'
        responses:
          '201':
            description: User created successfully
          '400':
            description: Request has not the wantet format
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '500':
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
      get:
        tags:
          - Users
        summary: Gets all users
        description: |
          Delivers all users only if the user is admin
        responses:
          '200':
            description: Users
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/User'
          '400':
            description: Request has not the needed format
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '401':
            description: No Token
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '403':
            description: Not the needed permission
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '500':
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /users/{username}:
    get:
      tags:
        - Users
      summary: Get the data of the User
      description: |
        Delivers all data of the User with the given username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Users
      summary: Deletes one User
      description: |
        Deletes one User with the given username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
      responses:
        '204':
          description: Successfully deleted
        '400':
          description: The username does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Users
      summary: Overwrites a User
      description: |
        Overwrites a User with the given username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
        description: User to overwrite (username must match username in path)
      responses:
        '201':
          description: User overwritten
        '400':
          description: Username does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/vehicles:
    post:
      tags:
        - Vehicles
      summary: Add a new Vehicle to the user
      description: |
        Inserts a new Vehicle for the given User into the database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        '201':
          description: Vehicle created succesfully
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Vehicles
      summary: Get the vehicles of the user
      description: |
        Delivers all vehicles of the given username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
      responses:
        '200':
          description: Vehicles of the User
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{username}/vehicles/{vin}/singleCosts:
    post:
      tags:
        - Single Costs
      summary: Add a new single cost item to the Vehicle
      description: |
        Inserts a new single cost item for the given Vehicle into the database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleCostItem'
      responses:
        '201':
          description: Single cost item created succesfully
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Single Costs
      summary: Get the single cost item of the vehicle
      description: |
        Delivers all single cost items of the given vehicle
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      responses:
        '200':
          description: Single cost items of the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SingleCostItemComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/vehicles/{vin}/singleCosts/{id}:
    get:
      tags:
        - Single Costs
      summary: Get a specific single cost item of the vehicle
      description: |
        Delivers a specific single cost item of the given id with the givenvin and username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
        - in: path
          name: id
          schema:
            type: string
            format: uuid
            example: '56b31f28-6bb2-48dc-b93d-2b6ef0377410'
          required: true
          description: ID of the searche Cost Item
      responses:
        '200':
          description: Single cost item of the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SingleCostItemComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user or vin or id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Single Costs
      summary: Deletes one singel cost item
      description: |
        Deletes one singel cost item with the given vin
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
        - in: path
          name: id
          schema:
            type: string
            format: uuid
            example: '56b31f28-6bb2-48dc-b93d-2b6ef0377410'
          required: true
          description: ID of the searche Cost Item
      responses:
        '204':
          description: Successfully deleted
        '400':
          description: The id does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user or vin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Single Costs
      summary: Overwrites a single cost item
      description: |
        Overwrites a single cost item
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
        - in: path
          name: id
          schema:
            type: string
            format: uuid
            example: '56b31f28-6bb2-48dc-b93d-2b6ef0377410'
          required: true
          description: ID of the searche Cost Item
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleCostItem'
        description: Single cost item to overwrite (vin must match vin in path)
      responses:
        '201':
          description: ID overwritten
        '400':
          description: The Id does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user or vin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      required:
        - username
        - mail
        - password
        - role
      properties:
        username:
          type: string
          example: "Mustermann25"
        mail:
          type: string
          example: "max.mustermann@host.de"
        password:
          type: string
          example: "Test-123"
        role:
          type: string
          example: admin | member
    UserInput:
      type: object
      required:
        - username
        - mail
        - password
        - passwordRepeat
        - role
      properties:
        username:
          type: string
          example: "Mustermann25"
        mail:
          type: string
          example: "max.mustermann@host.de"
        password:
          type: string
          example: "Test-123"
        passwordRepeat:
          type: string
          example: "Test-123"
        role:
          type: string
          example: admin | member
    Vehicle:
      type: object
      required:
        - vin
      properties:
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        year:
          type: integer
          format: int32
          example: 2001
        make:
          type: "string"
          example: "BMW"
        model:
          type: "string"
          example: "116i"
        type:
          type: "string"
          example: "Small car"
        color:
          type: "string"
          example: "Blue"
        weight:
          type: integer
          format: int32
          example: 2862
        license:
          type: string
          example: "Tü-T-7502"
    VehicleComplete:
      type: object
      required:
        - vin
        - username
      properties:
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        username:
          type: string
          example: "Mustermann25"
        year:
          type: string
          format: Date
          example: "Fri Dec 03 2010 19:26:05 GMT+0100 (Mitteleuropäische Normalzeit)"
        make:
          type: "string"
          example: "BMW"
        model:
          type: "string"
          example: "116i"
        type:
          type: "string"
          example: "Small car"
        color:
          type: "string"
          example: "Blue"
        weight:
          type: integer
          format: int32
          example: 2862
        license:
          type: string
          example: "Tü-T-7502"
...
    SingleCostItem:
      type: object
      required:
        - date
        - price
        - type
        - name
      properties:
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        type:
          type: string
          description: type of the cost item. repair | procurement | rest
          example: "repair"
        name:
          type: string
          description: name of the cost item
          example: "new weels"
        description:
          type: string
          example: "weels for the winter season"
          description: closer description for the cost item (optional)
    SingleCostItemComplete:
      type: object
      required:
        - id
        - vin
        - username
        - date
        - price
        - type
        - name
      properties:
        id:
          type: string
          format: uuid
          example: "6c84fb90-12c4-11e1-840d-7b25c5ee775a"
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        username:
          type: string
          example: "Mustermann25"
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        type:
          type: string
          description: type of the cost item. repair | procurement | rest
          example: "repair"
        name:
          type: string
          description: name of the cost item
          example: "new weels"
        description:
          type: string
          example: "weels for the winter season"
          description: closer description for the cost item (optional)
...
    Error:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            type: object
            required:
              - msg
              - location
            properties:
              value:
                type: string
                example: TestValue
              msg:
                type: string
                example: Invalid Value
              param:
                type: string
                example: Password
              location:
                type: string
                example: body